python-sdk: Move exceptions to a single file
