--- .env.template
@@ -1,1 +1,0 @@
-FLAGSMITH_ENVIRONMENT_KEY=""
