--- src/exception/type_mismatch_error.py
@@ -1,8 +1,0 @@
-from src.exception.open_feature_error import OpenFeatureError
-from src.flag_evaluation.error_code import ErrorCode
-
-
-class TypeMismatchError(OpenFeatureError):
-    def __init__(self, error_message: str = None):
-        self.error_message = error_message
-        self.error_code = ErrorCode.TYPE_MISMATCH
